fix(cli): remove --no-prose, make build-theme tests self-build core
With the in-CLI fallback generator removed, the build-theme subprocess tests
require a compiled @xds/core (xds theme build imports @xds/core/theme, which
now hard-fails with ERR_CORE_NOT_FOUND if core isn't built). The CI test job
ran `pnpm test` without building core, so these tests started failing once
the fallback that previously masked the missing core was gone.

- build-theme tests: build @xds/core in beforeAll if dist is absent, so the
  subprocess `xds theme build` always has a compiled core — self-sufficient
  regardless of CI job ordering. Mirrors scripts/build-css.test.mjs.
- Remove the `theme build --no-prose` flag: the <Theme> runtime always emits
  prose defaults, so letting the build omit them was another CLI/runtime
  divergence. Prose now always ships in @layer reset. Rename the stale
  no-prose test to build-theme.prose.test.mjs.

NOTE: the test CI job would ideally also build @xds/core before `pnpm test`
(like the docsite-test job already does). That .github/workflows/ci.yml change
is omitted here because the push token lacks the `workflow` scope; the
beforeAll self-build makes the suite pass without it. A maintainer should add
the Build-core step to the test job.
